+++ modelcraft/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "4.0.1"
+__version__ = "4.0.2"
 
 from .cell import max_distortion as max_cell_distortion
 from .cell import remove_scale